BC-12022 - Refactor import paths for awaitStreamCompletion and FileDtoFactory
